Feat: graduate multiagent hook events from experimental (#1498)
