+++ src/strands/hooks/__init__.py
@@ -32,12 +32,18 @@ type-safe system that supports multiple subscribers per event type.
 from .events import (
     AfterInvocationEvent,
     AfterModelCallEvent,
+    # Multiagent hook events
+    AfterMultiAgentInvocationEvent,
+    AfterNodeCallEvent,
     AfterToolCallEvent,
     AgentInitializedEvent,
     BeforeInvocationEvent,
     BeforeModelCallEvent,
+    BeforeMultiAgentInvocationEvent,
+    BeforeNodeCallEvent,
     BeforeToolCallEvent,
     MessageAddedEvent,
+    MultiAgentInitializedEvent,
 )
 from .registry import BaseHookEvent, HookCallback, HookEvent, HookProvider, HookRegistry
 
@@ -56,4 +62,9 @@ __all__ = [
     "HookRegistry",
     "HookEvent",
     "BaseHookEvent",
+    "AfterMultiAgentInvocationEvent",
+    "AfterNodeCallEvent",
+    "BeforeMultiAgentInvocationEvent",
+    "BeforeNodeCallEvent",
+    "MultiAgentInitializedEvent",
 ]
