+++ src/strands/session/session_manager.py
@@ -9,12 +9,14 @@ from ..experimental.hooks.events import (
     BidiAgentInitializedEvent,
     BidiMessageAddedEvent,
 )
-from ..experimental.hooks.multiagent.events import (
+from ..hooks.events import (
+    AfterInvocationEvent,
     AfterMultiAgentInvocationEvent,
     AfterNodeCallEvent,
+    AgentInitializedEvent,
+    MessageAddedEvent,
     MultiAgentInitializedEvent,
 )
-from ..hooks.events import AfterInvocationEvent, AgentInitializedEvent, MessageAddedEvent
 from ..hooks.registry import HookProvider, HookRegistry
 from ..types.content import Message
