+++ tests_integ/hooks/multiagent/test_cancel.py
@@ -1,8 +1,7 @@
 import pytest
 
 from strands import Agent
-from strands.experimental.hooks.multiagent import BeforeNodeCallEvent
-from strands.hooks import HookProvider
+from strands.hooks import BeforeNodeCallEvent, HookProvider
 from strands.multiagent import GraphBuilder, Swarm
 from strands.multiagent.base import Status
 from strands.types._events import MultiAgentNodeCancelEvent
