+++ tests_integ/interrupts/multiagent/test_session.py
@@ -4,8 +4,7 @@ from unittest.mock import ANY
 import pytest
 
 from strands import Agent, tool
-from strands.experimental.hooks.multiagent import BeforeNodeCallEvent
-from strands.hooks import HookProvider
+from strands.hooks import BeforeNodeCallEvent, HookProvider
 from strands.interrupt import Interrupt
 from strands.multiagent import GraphBuilder, Swarm
 from strands.multiagent.base import Status
